refactor: userAdapter Dto 패키지 리팩터링
