+++ src/main/java/com/nhnacademy/frontend/auth/controller/LoginController.java
@@ -1,10 +1,10 @@
 package com.nhnacademy.frontend.auth.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nhnacademy.frontend.auth.domain.response.AdditionalSignupRequiredDto;
-import com.nhnacademy.frontend.auth.domain.response.OAuth2LoginResponseDto;
-import com.nhnacademy.frontend.auth.domain.response.PaycoCallbackResponseDto;
-import com.nhnacademy.frontend.auth.domain.response.ResponseDto;
+import com.nhnacademy.frontend.auth.dto.response.AdditionalSignupRequiredDto;
+import com.nhnacademy.frontend.auth.dto.response.OAuth2LoginResponseDto;
+import com.nhnacademy.frontend.auth.dto.response.PaycoCallbackResponseDto;
+import com.nhnacademy.frontend.auth.dto.response.ResponseDto;
 import com.nhnacademy.frontend.auth.service.AuthService;
 import com.nhnacademy.frontend.auth.util.JwtCookieUtil;
 import jakarta.servlet.http.Cookie;
