+++ src/main/java/com/nhnacademy/frontend/auth/dto/request/PasswordVerificationRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.auth.domain.request;
+package com.nhnacademy.frontend.auth.dto.request;
 
 public record PasswordVerificationRequestDto (
         String password
