+++ src/main/java/com/nhnacademy/frontend/auth/dto/request/UserIdCheckRequestDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.auth.domain.request;
+package com.nhnacademy.frontend.auth.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
