+++ src/main/java/com/nhnacademy/frontend/auth/dto/response/LoginResponseDto.java
@@ -1,11 +1,11 @@
-package com.nhnacademy.frontend.auth.domain.response;
+package com.nhnacademy.frontend.auth.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class RefreshTokenResponseDto {
+public class LoginResponseDto {
     private String accessToken;
     private String refreshToken;
 }
