+++ src/main/java/com/nhnacademy/frontend/auth/dto/response/RefreshTokenResponseDto.java
@@ -1,11 +1,11 @@
-package com.nhnacademy.frontend.auth.domain.response;
+package com.nhnacademy.frontend.auth.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class LoginResponseDto {
+public class RefreshTokenResponseDto {
     private String accessToken;
     private String refreshToken;
 }
