+++ src/main/java/com/nhnacademy/frontend/auth/service/impl/SignupServiceImpl.java
@@ -2,7 +2,7 @@ package com.nhnacademy.frontend.auth.service.impl;
 
 import com.nhnacademy.frontend.auth.service.SignupService;
 import com.nhnacademy.frontend.common.adapter.UserAdapter;
-import com.nhnacademy.frontend.auth.domain.request.UserCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.UserCreateRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
