+++ src/main/java/com/nhnacademy/frontend/common/adapter/UserAdapter.java
@@ -1,12 +1,12 @@
 package com.nhnacademy.frontend.common.adapter;
 
-import com.nhnacademy.frontend.admin.domain.response.ResponsePoint;
-import com.nhnacademy.frontend.auth.domain.request.UserCreateRequestDto;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponseAddress;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponsePointType;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponseUser;
-import com.nhnacademy.frontend.mypage.domain.request.AddressCreateRequest;
-import com.nhnacademy.frontend.mypage.domain.request.UserUpdateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponsePoint;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.UserCreateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponseAddress;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponsePointType;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponseUser;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.AddressCreateRequest;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.UserUpdateRequestDto;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -51,7 +51,6 @@ public interface UserAdapter {
     ResponseEntity<Page<ResponsePointType>> getPointTypeByGradeName(
             @RequestParam(name = "gradeName", required = false) String gradeName,
             Pageable pageable);
-
 
     @PutMapping("/user-api/users/bulk/grade")
     ResponseEntity<Void> bulkUpdateUserGrades();
