+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/user/request/AddressCreateRequest.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.mypage.domain.request;
+package com.nhnacademy.frontend.common.adapter.dto.user.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
