+++ src/main/java/com/nhnacademy/frontend/common/adapter/dto/user/response/ResponsePointType.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.frontend.common.adapter.domain.response;
+package com.nhnacademy.frontend.common.adapter.dto.user.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
