+++ src/main/java/com/nhnacademy/frontend/mypage/controller/MypageController.java
@@ -2,11 +2,11 @@ package com.nhnacademy.frontend.mypage.controller;
 
 
 import com.nhnacademy.frontend.auth.util.JwtCookieUtil;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponsePointType;
-import com.nhnacademy.frontend.admin.domain.response.ResponsePoint;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponseUser;
-import com.nhnacademy.frontend.mypage.domain.request.AddressCreateRequest;
-import com.nhnacademy.frontend.mypage.domain.request.UserUpdateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponsePointType;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponsePoint;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponseUser;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.AddressCreateRequest;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.UserUpdateRequestDto;
 import com.nhnacademy.frontend.mypage.service.MypageService;
 import feign.FeignException;
 import jakarta.servlet.http.HttpServletResponse;
