+++ src/test/java/com/nhnacademy/frontend/auth/filter/LoginFilterTest.java
@@ -1,7 +1,7 @@
 package com.nhnacademy.frontend.auth.filter;
 
-import com.nhnacademy.frontend.auth.domain.response.LoginResponseDto;
-import com.nhnacademy.frontend.auth.domain.response.TokenParseResponseDto;
+import com.nhnacademy.frontend.auth.dto.response.LoginResponseDto;
+import com.nhnacademy.frontend.auth.dto.response.TokenParseResponseDto;
 import com.nhnacademy.frontend.auth.service.AuthService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
