+++ src/test/java/com/nhnacademy/frontend/auth/handler/LoginSuccessHandlerTest.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.frontend.auth.handler;
 
-import com.nhnacademy.frontend.auth.domain.response.LoginResponseDto;
+import com.nhnacademy.frontend.auth.dto.response.LoginResponseDto;
 import com.nhnacademy.frontend.auth.util.JwtCookieUtil;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
