+++ src/test/java/com/nhnacademy/frontend/mypage/service/MypageServiceImplTest.java
@@ -1,14 +1,14 @@
 package com.nhnacademy.frontend.mypage.service;
 
-import com.nhnacademy.frontend.admin.domain.response.ResponsePoint;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponsePoint;
 import com.nhnacademy.frontend.auth.adapter.AuthAdapter;
-import com.nhnacademy.frontend.auth.domain.request.PasswordVerificationRequestDto;
+import com.nhnacademy.frontend.auth.dto.request.PasswordVerificationRequestDto;
 import com.nhnacademy.frontend.common.adapter.UserAdapter;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponseAddress;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponsePointType;
-import com.nhnacademy.frontend.common.adapter.domain.response.ResponseUser;
-import com.nhnacademy.frontend.mypage.domain.request.AddressCreateRequest;
-import com.nhnacademy.frontend.mypage.domain.request.UserUpdateRequestDto;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponseAddress;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponsePointType;
+import com.nhnacademy.frontend.common.adapter.dto.user.response.ResponseUser;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.AddressCreateRequest;
+import com.nhnacademy.frontend.common.adapter.dto.user.request.UserUpdateRequestDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
